requirements.txt
# Core API dependencies
fastapi==0.116.1
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.7

# Authentication
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0

# Database
sqlalchemy==2.0.43
alembic==1.16.5
asyncpg==0.29.0
geoalchemy2==0.14.3

# Redis
redis==5.0.1
hiredis==2.3.2

# Cloud/Storage
boto3==1.34.34
python-dotenv==1.0.1

# Utilities
httpx==0.26.0
aiofiles==23.2.1
pillow==10.3.0
structlog==24.1.0
email-validator==2.1.0.post1
itsdangerous==2.1.2
tenacity==8.2.3

# Geospatial
geopy==2.4.1
numpy<2.0
shapely==2.0.3
geohash2==1.1

# Worker/Queue
celery==5.3.4

# Data generation
faker==25.2.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
fakeredis==2.21.1
httpx==0.26.0

# Development
black==23.12.1

# Monitoring
prometheus-client==0.19.0